requirements.txt
absl==0.0
absl_py==0.12.0
Keras==2.4.3
matplotlib==3.5.1
numpy==1.20.2
object_detection==0.0.3
pandas==1.4.2
Pillow==9.1.0
simple_image_download==0.5
tensorflow==2.8.0
